Monitoring_Price/telegram_bot/requirements.txt
python-dotenv
aiogram==2.25.1
requests
matplotlib
pydantic
pydantic_settings
pandas